[fixed] Unmounting players on incap and death [fixed] Unmounting now takes place before „clearStates()“ [fixed] Using @kill while dueling is not allowed now - this was causing a deadlock!
Due to the container/parent changes, players are technically in a container now when mounted on a vehicle/pet – the mount is the container.

If a player gets killed on/in a mount, the parentID for the player (the vehicle) wasn't reset.
If the player logs back in later and the vehicle isn't existing (called) anymore, the player gets stucked in the loading screen.

This patch makes sure, that a player gets dismounted on death.

Code-fix:
clearStates() is clearing the mount state as well, so bool->isMounted() was cleared before the player got unmounted.



git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1184 c3d1530f-68f5-4bd0-87dc-8ef779617e40